# [inductor][Gemm] Autotune with matrix_instr_nonkdim for AMDGPU