@jataylo 2024-03-27
Hi @htyu @xw285cornell, great work integrating this which seems to help the config performance, but I suspect this change may be causing an unintended bug.

It seems that `matrix_instr_nonkdim` is used for the autotune benchmark as it is provided in do_bench args and improves the overall performance.

But it does not seem to be used in the generated wrapper code itself, which can degrade the model performance. It seems that we would require a change to the triton_heuristic.template decorator to support this setting
`https://github.com/htyu/pytorch/blob/72ca297c609c6d149a77ae256320f7fc5685758e/torch/_inductor/select_algorithm.py#L158`

=============================================================================

Example:
https://gist.github.com/jataylo/488e32f304958f4a220913ff6afb268f

The above gist is from an addmm autotune, which arrives at the optimal config of `num_stages=1, warp_size=8, matrix_instr=16` for this case, but we see that switching to num_stages=0 actually outperforms the autotuned config even though it was not the optimal in autotuning

  | addmm_example (ms)
-- | --
num_stages=0 | 0.166
num_stages=1 | 0.207

To ensure `matrix_instr_nonkdim` is indeed being used I modified `triton_meta` (https://gist.github.com/jataylo/488e32f304958f4a220913ff6afb268f#file-addmm_wrapper-py-L20) to include matrix_instr_nonkdim setting which shows that the autotuned config with matrix_instr_nonkdim=16 is the fastest:
  | addmm_example (ms)
-- | --
num_stages=0,   matrix_instr_nonkdim=0 | 0.166
num_stages=0,   matrix_instr_nonkdim=16 | 0.14
num_stages=1,   matrix_instr_nonkdim=0 | 0.207
**num_stages=1,   matrix_instr_nonkdim=16 | 0.134**