@htyu 2024-03-27
> https://gist.github.com/jataylo/488e32f304958f4a220913ff6afb268f#file-addmm_wrapper-py-L20

Thanks for the detailed analysis! A fix is indeed needed as you pointed out. Would you like me to work on the fix, or you'd like to send out a PR for this since you are already on it?