@htyu 2024-03-28
Fixed by https://github.com/pytorch/pytorch/pull/122852